refactor(intelligence) Restructure remote engine
